--- a/frontend/app.tsx
+++ b/frontend/app.tsx
@@ -1,7 +1,6 @@
-import React, { memo, useCallback, useEffect, useReducer } from "react";
+import React, { memo, useCallback, useEffect } from "react";
 import type {
   ExperimentalDiff as Diff,
-  ReadonlyJSONValue,
   ReadTransaction,
   Replicache,
 } from "replicache";
@@ -30,7 +29,7 @@ import TopFilter from "./top-filter";
 import IssueList from "./issue-list";
 import { useQueryState } from "next-usequerystate";
 import IssueBoard from "./issue-board";
-import { isEqual, minBy, partial, pickBy, sortBy, sortedIndexBy } from "lodash";
+import { isEqual, minBy, pickBy } from "lodash";
 import IssueDetail from "./issue-detail";
 import { generateKeyBetween } from "fractional-indexing";
 import { useSubscribe } from "replicache-react";
@@ -38,7 +37,10 @@ import classnames from "classnames";
 import { getPartialSyncState, PartialSyncState } from "./control";
 import type { UndoManager } from "@rocicorp/undo";
 import { HotKeys } from "react-hotkeys";
+import { Materialite, PersistentTreeView } from "@vlcn.io/materialite";
+import type { PersistentSetSource } from "@vlcn.io/materialite";
 
+const materialite = new Materialite();
 class Filters {
   private readonly _viewStatuses: Set<Status> | undefined;
   private readonly _issuesStatuses: Set<Status> | undefined;
@@ -96,11 +98,11 @@ class Filters {
     }
   }
 
-  viewFilter(issue: Issue): boolean {
+  viewFilter = (issue: Issue): boolean => {
     return this._viewStatuses ? this._viewStatuses.has(issue.status) : true;
-  }
+  };
 
-  issuesFilter(issue: Issue): boolean {
+  issuesFilter = (issue: Issue): boolean => {
     if (this._issuesStatuses) {
       if (!this._issuesStatuses.has(issue.status)) {
         return false;
@@ -112,7 +114,7 @@ class Filters {
       }
     }
     return true;
-  }
+  };
 
   equals(other: Filters): boolean {
     return (
@@ -154,35 +156,6 @@ function getTitle(view: string | null) {
   }
 }
 
-type State = {
-  allIssuesMap: Map<string, Issue>;
-  viewIssueCount: number;
-  filteredIssues: Issue[];
-  filters: Filters;
-  issueOrder: Order;
-};
-function timedReducer(
-  state: State,
-  action:
-    | {
-        type: "diff";
-        diff: Diff;
-      }
-    | {
-        type: "setFilters";
-        filters: Filters;
-      }
-    | {
-        type: "setIssueOrder";
-        issueOrder: Order;
-      }
-): State {
-  const start = Date.now();
-  const result = reducer(state, action);
-  console.log(`Reducer took ${Date.now() - start}ms`, action);
-  return result;
-}
-
 function getOrderValue(issueOrder: Order, issue: Issue): string {
   let orderValue: string;
   switch (issueOrder) {
@@ -211,130 +184,48 @@ function getOrderValue(issueOrder: Order, issue: Issue): string {
   return orderValue;
 }
 
-function reducer(
-  state: State,
-  action:
-    | {
-        type: "diff";
-        diff: Diff;
-      }
-    | {
-        type: "setFilters";
-        filters: Filters;
-      }
-    | {
-        type: "setIssueOrder";
-        issueOrder: Order;
-      }
-): State {
-  const filters = action.type === "setFilters" ? action.filters : state.filters;
-  const issueOrder =
-    action.type === "setIssueOrder" ? action.issueOrder : state.issueOrder;
-  const orderIteratee = partial(getOrderValue, issueOrder);
-  function filterAndSort(issues: Issue[]): Issue[] {
-    return sortBy(
-      issues.filter((issue) => filters.issuesFilter(issue)),
-      orderIteratee
-    );
-  }
-  function countViewIssues(issues: Issue[]): number {
-    let count = 0;
-    for (const issue of issues) {
-      if (filters.viewFilter(issue)) {
-        count++;
-      }
-    }
-    return count;
-  }
-
-  switch (action.type) {
-    case "diff": {
-      return diffReducer(state, action.diff);
-    }
-    case "setFilters": {
-      if (action.filters.equals(state.filters)) {
-        return state;
-      }
-      const allIssues = [...state.allIssuesMap.values()];
-      return {
-        ...state,
-        viewIssueCount: countViewIssues(allIssues),
-        filters: action.filters,
-        filteredIssues: filterAndSort(allIssues),
-      };
-    }
-    case "setIssueOrder": {
-      if (action.issueOrder === state.issueOrder) {
-        return state;
-      }
-      return {
-        ...state,
-        filteredIssues: sortBy(state.filteredIssues, orderIteratee),
-        issueOrder: action.issueOrder,
-      };
-    }
-  }
-
-  return state;
+function issueCountView(
+  source: PersistentSetSource<Issue>,
+  filter: (i: Issue) => boolean
+) {
+  return source.stream.filter(filter).size().materializePrimitive(0);
 }
 
-function diffReducer(state: State, diff: Diff): State {
-  if (diff.length === 0) {
-    return state;
-  }
-  const newAllIssuesMap = new Map(state.allIssuesMap);
-  let newViewIssueCount = state.viewIssueCount;
-  const newFilteredIssues = [...state.filteredIssues];
-  const orderIteratee = partial(getOrderValue, state.issueOrder);
+function filteredIssuesView(
+  source: PersistentSetSource<Issue>,
+  order: Order,
+  filter: (i: Issue) => boolean
+) {
+  return source.stream
+    .filter(filter)
+    .materialize((l: Issue, r: Issue) =>
+      getOrderValue(order, l).localeCompare(getOrderValue(order, r))
+    );
+}
 
-  function add(key: string, newValue: ReadonlyJSONValue) {
-    const newIssue = issueFromKeyAndValue(key, newValue);
-    newAllIssuesMap.set(key, newIssue);
-    if (state.filters.viewFilter(newIssue)) {
-      newViewIssueCount++;
-    }
-    if (state.filters.issuesFilter(newIssue)) {
-      newFilteredIssues.splice(
-        sortedIndexBy(newFilteredIssues, newIssue, orderIteratee),
-        0,
-        newIssue
-      );
-    }
+function onNewDiff(diff: Diff) {
+  if (diff.length === 0) {
+    return;
   }
-  function del(key: string, oldValue: ReadonlyJSONValue) {
-    const oldIssue = issueFromKeyAndValue(key, oldValue);
-    const index = sortedIndexBy(newFilteredIssues, oldIssue, orderIteratee);
-    newAllIssuesMap.delete(key);
-    if (state.filters.viewFilter(oldIssue)) {
-      newViewIssueCount--;
-    }
-    if (newFilteredIssues[index]?.id === oldIssue.id) {
-      newFilteredIssues.splice(index, 1);
-    }
-  }
-  for (const diffOp of diff) {
-    switch (diffOp.op) {
-      case "add": {
-        add(diffOp.key as string, diffOp.newValue);
-        break;
+
+  const start = performance.now();
+  materialite.tx(() => {
+    for (const diffOp of diff) {
+      if ("oldValue" in diffOp) {
+        allIssueSet.delete(
+          issueFromKeyAndValue(diffOp.key as string, diffOp.oldValue)
+        );
       }
-      case "del": {
-        del(diffOp.key as string, diffOp.oldValue);
-        break;
-      }
-      case "change": {
-        del(diffOp.key as string, diffOp.oldValue);
-        add(diffOp.key as string, diffOp.newValue);
-        break;
+      if ("newValue" in diffOp) {
+        allIssueSet.add(
+          issueFromKeyAndValue(diffOp.key as string, diffOp.newValue)
+        );
       }
     }
-  }
-  return {
-    ...state,
-    allIssuesMap: newAllIssuesMap,
-    viewIssueCount: newViewIssueCount,
-    filteredIssues: newFilteredIssues,
-  };
+  });
+
+  const duration = performance.now() - start;
+  console.log(`Diff duration: ${duration}`);
 }
 
 type AppProps = {
@@ -342,6 +233,16 @@ type AppProps = {
   undoManager: UndoManager;
 };
 
+const allIssueSet = materialite.newPersistentSet<Issue>((l, r) =>
+  l.id.localeCompare(r.id)
+);
+type IssueViews = {
+  allIssues: PersistentSetSource<Issue>["data"];
+  filteredIssues: PersistentTreeView<Issue>["data"];
+  issueCount: number;
+  hasNonViewFilters: boolean;
+  // minKanban: Issue;
+};
 const App = ({ rep, undoManager }: AppProps) => {
   const [view] = useQueryState("view");
   const [priorityFilter] = useQueryState("priorityFilter");
@@ -350,13 +251,54 @@ const App = ({ rep, undoManager }: AppProps) => {
   const [detailIssueID, setDetailIssueID] = useQueryState("iss");
   const [menuVisible, setMenuVisible] = useState(false);
 
-  const [state, dispatch] = useReducer(timedReducer, {
-    allIssuesMap: new Map(),
-    viewIssueCount: 0,
-    filteredIssues: [],
-    filters: getFilters(view, priorityFilter, statusFilter),
-    issueOrder: getIssueOrder(view, orderBy),
+  const [issueViews, setIssueViews] = useState<IssueViews>({
+    allIssues: allIssueSet.data,
+    issueCount: 0,
+    filteredIssues: allIssueSet.data,
+    hasNonViewFilters: false,
   });
+
+  useEffect(() => {
+    const filters = getFilters(view, priorityFilter, statusFilter);
+    const order = getIssueOrder(view, orderBy);
+    const filterView = filteredIssuesView(
+      allIssueSet,
+      order,
+      filters.issuesFilter
+    );
+    const countView = issueCountView(allIssueSet, filters.viewFilter);
+    const removeFilterListener = filterView.onChange((data) => {
+      setIssueViews((last) => ({
+        ...last,
+        allIssues: allIssueSet.data,
+        filteredIssues: data,
+        hasNonViewFilters: filters.hasNonViewFilters,
+      }));
+    });
+    const removeCountListener = countView.onChange((data) => {
+      setIssueViews((last) => ({
+        ...last,
+        allIssues: allIssueSet.data,
+        issueCount: data,
+        hasNonViewFilters: filters.hasNonViewFilters,
+      }));
+    });
+    const removeAllIssueListener = allIssueSet.onChange((data) => {
+      setIssueViews((last) => ({
+        ...last,
+        allIssues: data,
+        hasNonViewFilters: filters.hasNonViewFilters,
+      }));
+    });
+    // TODO (mlaw): remove the need for this call.
+    // The framework knows when new views are attached and when this is needed.
+    allIssueSet.recomputeAll();
+    return () => {
+      removeFilterListener();
+      removeCountListener();
+      removeAllIssueListener();
+    };
+  }, [priorityFilter, statusFilter, orderBy, view]);
 
   const partialSync = useSubscribe<
     PartialSyncState | "NOT_RECEIVED_FROM_SERVER"
@@ -377,35 +319,18 @@ const App = ({ rep, undoManager }: AppProps) => {
   }, [rep, partialSync, partialSyncComplete]);
 
   useEffect(() => {
-    rep.experimentalWatch(
-      (diff) => {
-        dispatch({
-          type: "diff",
-          diff,
-        });
-      },
-      { prefix: ISSUE_KEY_PREFIX, initialValuesInFirstDiff: true }
-    );
+    return rep.experimentalWatch(onNewDiff, {
+      prefix: ISSUE_KEY_PREFIX,
+      initialValuesInFirstDiff: true,
+    });
   }, [rep]);
-
-  useEffect(() => {
-    dispatch({
-      type: "setFilters",
-      filters: getFilters(view, priorityFilter, statusFilter),
-    });
-  }, [view, priorityFilter, statusFilter]);
-
-  useEffect(() => {
-    dispatch({
-      type: "setIssueOrder",
-      issueOrder: getIssueOrder(view, orderBy),
-    });
-  }, [view, orderBy]);
 
   const handleCreateIssue = useCallback(
     async (issue: Omit<Issue, "kanbanOrder">, description: Description) => {
-      const minKanbanOrderIssue = minBy(
-        [...state.allIssuesMap.values()],
+      // TODO (mlaw): this kanban thing
+      const minKanbanOrderIssue = minBy<Issue>(
+        // [...state.allIssuesMap.values()],
+        [],
         (issue) => issue.kanbanOrder
       );
       const minKanbanOrder = minKanbanOrderIssue
@@ -420,7 +345,7 @@ const App = ({ rep, undoManager }: AppProps) => {
         description,
       });
     },
-    [rep.mutate, state.allIssuesMap]
+    [rep.mutate /*state.allIssuesMap*/]
   );
   const handleCreateComment = useCallback(
     async (comment: Comment) => {
@@ -508,7 +433,7 @@ const App = ({ rep, undoManager }: AppProps) => {
         view={view}
         detailIssueID={detailIssueID}
         isLoading={!partialSyncComplete}
-        state={state}
+        state={issueViews}
         rep={rep}
         onCloseMenu={handleCloseMenu}
         onToggleMenu={handleToggleMenu}
@@ -531,7 +456,7 @@ interface LayoutProps {
   view: string | null;
   detailIssueID: string | null;
   isLoading: boolean;
-  state: State;
+  state: IssueViews;
   rep: Replicache<M>;
   onCloseMenu: () => void;
   onToggleMenu: () => void;
@@ -576,11 +501,9 @@ const RawLayout = ({
               onToggleMenu={onToggleMenu}
               title={getTitle(view)}
               filteredIssuesCount={
-                state.filters.hasNonViewFilters
-                  ? state.filteredIssues.length
-                  : undefined
+                state.hasNonViewFilters ? state.filteredIssues.size : undefined
               }
-              issuesCount={state.viewIssueCount}
+              issuesCount={state.issueCount}
               showSortOrderMenu={view !== "board"}
             />
           </div>
@@ -624,5 +547,4 @@ const RawLayout = ({
 };
 
 const Layout = memo(RawLayout);
-
 export default App;