--- a/package.json
+++ b/package.json
@@ -41,7 +41,7 @@
     "replicache": ">=13.0.0",
     "replicache-react": "3.0.0",
     "zod": "^3.13.4",
-    "@vlcn.io/materialite": "file:../vulcan/materialite/public/packages/materialite"
+    "@vlcn.io/materialite": "^2.0.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^6.2.1",