Later PRs:

- incrementally maintain the kanban board. Working in: https://github.com/tantaman/repliear/pull/4
- ref count operators rather than destroying at source. For a future PR.

---
This PR:

Scaled to 10x the issues (113,070)

**The good:**
- updates take ~2-4ms on this branch vs ~20ms without these changes.
  - note: we can get this down to 1-2ms via a source which doesn't retain values. Materialite is saving values in the source (SetSource / MapSource) as well as the resulting view. Replicache already has the values so we could remove the extra copy we're maintaining in the source set.
- page refresh is 4x faster on this branch (0.5 seconds to load all issues vs 2 seconds)
- ~80 lines less code (package.lock.json accounts for 50 line additions in the diff)

**The meh:**
- Completely changing filters can be 3x slower. **Update: We can fix all of this by sorting the set _before_ materializing the set**
   - Depends on result size.
   - To go from a filtered set to showing all 100,000 results takes 300ms where as the original implementation takes 100ms. 
   - In a way this is expected -- updating an immutable tree with the entire result set is slower than copying an array of the entire result set. In a way it is unexpected since full page refresh is 4x faster on this branch.
   - The smaller the resulting filtered set, the smaller the perf difference.

Another good test would be typing and updating the issue table as the user types. Based on my other experiments, this works fine with materialite (we do very few allocations) where as it quickly crashes otherwise (too many array copies).

**Unrelated things I noticed:**
1. Two diffs are fired for each change event. The first changes the field and modified time, the second changes the modified time again without changing anything else.

![Screenshot 2023-10-25 at 9 39 50 AM](https://github.com/tantaman/repliear/assets/1009003/9bc09ae8-80d6-4c92-aef8-85286eb63c9c)

![Screenshot 2023-10-25 at 9 40 26 AM](https://github.com/tantaman/repliear/assets/1009003/b9f811bd-eb25-4225-a5be-50e6a0bd0e3e)

![Screenshot 2023-10-25 at 9 40 28 AM](https://github.com/tantaman/repliear/assets/1009003/e7cc80b6-7b0a-4c80-b046-71f1d6c0b7a0)


2. the `onSuccess` callback inside replicache that is invoked after a mutation often causes a bunch of frame drops

![Screenshot 2023-10-25 at 9 33 25 AM](https://github.com/tantaman/repliear/assets/1009003/fcb6bc50-4dbc-4e0f-b105-4d4478554f69)

3. After leaving `repliear` open a while (both on this branch and on main), it just starts frame dropping like crazy and all interactions are slow (even just moving the mouse and not clicking anything)

![Screenshot 2023-10-25 at 10 13 53 AM](https://github.com/tantaman/repliear/assets/1009003/e8bceea1-4daa-45cb-8a44-cb9b083ab668)

4. initial load time is also much faster on this branch. This is great but unexpected.

**Before (2+ seconds to refresh):**

https://github.com/tantaman/repliear/assets/1009003/df876de6-018e-471b-9119-787927e7b9a5

**After (0.5 second to refresh):**

https://github.com/tantaman/repliear/assets/1009003/c1923048-a5b0-4d32-9d8c-5710a7ddd54f