@tantaman 2023-11-01
closing this out.

Re-doing with some new APIs to deal with materialization, sorted sources and pagination.

https://github.com/vlcn-io/materialite/pull/1